% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_banzhaf.R
\name{gt_banzhaf}
\alias{gt_banzhaf}
\title{Banzhaf index}
\usage{
gt_banzhaf(w, q = 0.5)
}
\arguments{
\item{w}{Vector of voting weights}

\item{q}{Threshold for voting success}
}
\description{
Function for calculating the Banzhaf index.
Done by taking all coalitions and adding +1 for decisive coalitions, -1 otherwise
}
\examples{
 gt_banzhaf(w = 1:5, q=.5)
}
\keyword{Banzhaf,}
\keyword{Decision}
\keyword{Rules}
