% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_declarations.R
\name{gt_cascade}
\alias{gt_cascade}
\title{Calculated declarations from an Information Cascade}
\usage{
gt_cascade(x, signals = x)
}
\arguments{
\item{Signals}{Vector with signals received}
}
\description{
A  cascade  can  also  form,  for  example,  if  the  first  two  decisions  differ  and  the  next  two  match. In  all  cases,  it  takes  an  imbalance  of  two  de-  cisions  in  one  direction  to  overpower  the  informational  content  of  subsequent  individual  signals. PosteriorFunction to figure if you should say Yes or no given a string with n relevant Yeses and m relevant no's
}
\examples{
 gt_cascade(c(1,0,1,1,0))
 gt_cascade(c(0,0,1,1,0))
}
\keyword{Cascade}

