% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_cgm.R
\name{gt_cgm}
\alias{gt_cgm}
\title{Clarke-Groves Mechanism.}
\usage{
gt_cgm(select = function(i, j) (i + j)^2/4, theta1 = 0.6, theta2 = 0.3,
  normalize = TRUE, normalizevalue = 100, u = function(x, theta) theta *
  (x^0.5), cost = function(x) x, declarations1 = c(0, 0.3, 0.6, 1),
  declarations2 = declarations1, ucex = 1.5, h = function(j) -(j^2/2),
  scale = 1000, maincex = 2, col.main = "red", mainline = -1.7,
  n.grid = 500, clevels = 4, at_axis = c(0, 0.3, 0.6, 1),
  addpoint = FALSE, gcnash = TRUE, hack = FALSE)
}
\arguments{
\item{theta1}{Player 1's valuation}

\item{theta2}{Player 2's valuation}
}
\description{
Utility for public good is of the form u(x | theta_i) = theta_i*x^.5 - t_i. Default cost equals amount produced, say.  Consider two games. In one each is charged x/2. In the other VCG transfers are used.  Games solved as normal form games (as if in fact players know the other's preferences, even if mechanism does not)
}
\examples{
 gt_cgm()
}
\keyword{Clarke}
\keyword{Design}
\keyword{Groves}
\keyword{Mechanism}
\keyword{Mechanism,}

