% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{gt_curve}
\alias{gt_curve}
\title{Drawing a curve}
\usage{
gt_curve(cx = 0, cy = 0, radx = 1, rady = 1, col = "red",
  fine = 100, type = "l", from = 0, run = 1, new = FALSE, main = "",
  lwd = 1, arrow = FALSE, arlength = (radx + rady)/4, tilt = 0,
  lty = 1, vectors = FALSE)
}
\arguments{
\item{cx}{a number x center of curve}

\item{cy}{a number y center of curve}
}
\description{
Curve segment starting; let circle be indxed by 0,1, where 0 is the beginning of the circle (center right) moving clockwise to 1 back to center right
run is the length of the curve, so run = 1 is a full circle, run = .5 is a semi circle etc; run -x, goes anticlockwise
Then semi circle might start at .5 and end at 1, or start even at .75 and end at 2.
}
\examples{
gt_curve()
}
\keyword{arrows}
\keyword{curve}
