% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_cycles.R
\name{gt_cycles}
\alias{gt_cycles}
\title{Calculate and plot chaotic cycles}
\usage{
gt_cycles(n_voters = 3, n = n_voters, n_motions = 3, k = n_motions,
  seed = 1, z = 3, col1 = "red", coli = "black", colp = "red",
  pointcex = 0.5, xlab = "Issue 1", ylab = "Issue 2", ann = TRUE,
  axes = TRUE, a2z = FALSE, a2zcol = "red", a2zw = 2,
  a2zclose = 0.025, alab = "", zlab = "", inner = FALSE,
  colrange = rep(col1, k), col2 = colrange[k], mar = rep(2, 4))
}
\arguments{
\item{n}{Number of voters}

\item{k}{Number of motions}

\item{seed}{optionally set a seed}

\item{z}{size of plane for plotting xlim(-z,z)}

\item{a}{Alternative policy}
}
\description{
Plots chaotic majority rule cycles. We randomly select n voters and construct a sequence of k motions, each of which is majority preferred to the previous one in the sequence. Parameter z is used to scale the graph and a seed ("seed") is used for replication
}
\examples{
gt_cycles(3,5)


}
\keyword{Chaos}
\keyword{Plott,}
\keyword{Spatial,}
