% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_folkmm.R
\name{gt_folk}
\alias{gt_folk}
\title{Folk Theorem Illustration}
\usage{
gt_folk(X, Y = t(X), fine = 100, feasible = TRUE, rational = TRUE,
  pointsize = 1, colf = "grey", colr = rgb(1, 0.1, 0, 0.5), main = "",
  mainsize = 1, ...)
}
\arguments{
\item{X}{Payoff matrix for player 1. Defaults to coordination.}

\item{Y}{Payoff matrix for player 2. Defaults to coordination.}
}
\description{
Graphs the individually rational and feasible outcomes that can be sustained as equilibria of a two person game
}
\examples{
X = matrix(c(2,3,0,1),2)
gt_folk(X)
X = matrix(  c(2,0,0,0,1,0,0,0,5),3)
Y = t(matrix(c(1,2,1,2,1,1,1,3,1),3))
gt_folk(X,Y)

}
\keyword{Nash}
\keyword{Payoff}
\keyword{matrix,}
