% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_leg_barg.R
\name{gt_leg_barg}
\alias{gt_leg_barg}
\title{Legislative bargaining}
\usage{
gt_leg_barg(p, seed = NULL)
}
\arguments{
\item{p}{A vector of recognition probabilities}

\item{seed}{A seed}
}
\description{
Legislative bargaining with three equally weighted voters. For any recognition probabilities, return the equilibrium allocations. Note: when lowest unit has 0 probability then a random solution taken
}
\examples{
 gt_leg_barg(c(.7,.3,0))
}
\keyword{MIW}

