% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_spatial_preferences.R
\name{gt_maj_gain}
\alias{gt_maj_gain}
\title{Assesses how much  majorities gain from a policy change in spatial model}
\usage{
gt_maj_gain(P, a, b, method = "euclidean")
}
\arguments{
\item{P}{matrix of ideal points}

\item{a}{Alternative policy}

\item{b}{Status quo policy}

\item{method}{Distance metric, deafulats to Euclidean}
}
\description{
Assesses how much  majorities gain from a policy change in spatial model
}
\examples{
gt_maj_gain(matrix(c(0, 0, .5, 1, 1, 0), 2), 0:1, c(.5,.5))
}
\keyword{Chaos}
\keyword{Plott,}
\keyword{Spatial,}

