% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_majority_phase.R
\name{gt_majority_phase}
\alias{gt_majority_phase}
\title{Phase plot for majority rule. Requires majority preference relation function.}
\usage{
gt_majority_phase(P, n = 11, grid = 4, pointsize = 2, box = TRUE,
  raylengths = c(0.5), raycols = c("red", "pink"), rayarrs = c(0.15, 0.1),
  q = 0.5, pointcol = "grey", lwd = 1, main = "")
}
\arguments{
\item{P}{Matrix with location of ideal points}

\item{n}{number of arrow per region}

\item{grid}{number of grid points}

\item{raylengths}{vector with lengths of distances to examine}
}
\description{
Phase plot for majority rule. Requires majority preference relation function.
}
\examples{
gt_majority_phase(matrix(c(0, 0, .5, 1, 1, 0), 2), 31, raylengths=.5)
}
\keyword{Plott}

