% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_minimax.R
\name{gt_minimax}
\alias{gt_minimax}
\title{Minimax}
\usage{
gt_minimax(A, fine = 200)
}
\arguments{
\item{A}{Payoff matrix}
}
\description{
Calculates minimax strategies and payoffs on a probability grid with increments of 1/fine.  Matrix A is the matrix for a single player with own actions in the row position
}
\examples{
A <- matrix(c(2,0,0,0,1,0,0,0,5),3)
gt_minimax(A, 200)

}
\keyword{Minimax}

