% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{gt_permv}
\alias{gt_permv}
\title{Generates set of permutations. Credit: Bernd Beber}
\usage{
gt_permv(v)
}
\arguments{
\item{v}{a vector giving number of elements in each variable}
}
\description{
Generates set of permutations. Credit: Bernd Beber
}
\examples{
gt_permv(c(2,3))
}
\keyword{permutations}
