% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_plot_mr.R
\name{gt_plot_mr}
\alias{gt_plot_mr}
\title{Melzer Richard Illustration}
\usage{
gt_plot_mr(e, tax = seq(0, 1, 0.01), ymax = gt_mr_utility(gt_mr_maxu(e), e),
  main = "", xlab = "", scale = 1, ylab = "", color = "grey",
  gt_mr_distribution = function(tax, e) tax * (1 - tax)/2,
  gt_mr_production = function(tax, e) (1 - tax)^2 * e/4,
  gt_mr_utility = function(tax, e) (gt_mr_distribution(tax, e) +
  gt_mr_production(tax, e)), gt_mr_maxu = function(e) (1 - e)/(2 - e))
}
\arguments{
\item{e}{Endowment}

\item{gt_mr_distribution}{function showing distribution given taxes}

\item{gt_mr_production}{function giving production given taxes}

\item{gt_mr_utility}{function giving utility given taxes}

\item{gt_mr_maxu}{function giving max utility given endowment}

\item{t}{tax range}
}
\description{
gt_plot_mr plots returns to actors in a Melzer RIchard economy given an endowment e
}
\examples{
#Illustration of bargaining outcomes from possibility set
gt_plot_mr(e=.5, main = "Endowment = .5")
}
\keyword{Melzer}
\keyword{Richard,}
\keyword{Tax}

