% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_qrule.R
\name{gt_qrule}
\alias{gt_qrule}
\title{Define q Rule (strict inequality)}
\usage{
gt_qrule(x, w = rep(1/length(x), length(x)), q = 0.5)
}
\arguments{
\item{x}{Vector of indicators of coalition membership}

\item{w}{Vector of voting weights}

\item{q}{Threshold for voting success}
}
\description{
Define q Rule (strict inequality)
}
\examples{
gt_qrule(x=c(1,1,0,0,0), q=.5)
 gt_qrule(x=c(1,1,1,0,0), q=.5)
}
\keyword{Condorcet,}
\keyword{Decision}
\keyword{RUles}

