% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{gt_shadowtext}
\alias{gt_shadowtext}
\title{Shadow text}
\usage{
gt_shadowtext(x, y = NULL, labels, col = "black", bg = "white",
  theta = seq(pi/4, 2 * pi, length.out = 8), r = 0.1, ...)
}
\arguments{
\item{x}{X location}
}
\description{
Place a shadow behind text (white default)
}
\examples{
gt_shadowtext()
}
\keyword{text}

