% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{gt_shift}
\alias{gt_shift}
\title{Helper to shift coordinates}
\usage{
gt_shift(s, l, d)
}
\arguments{
\item{s}{a number -- center of shift}

\item{l}{a number -- length of shift}

\item{d}{a number -- size of shift}
}
\description{
Generate a sequence centered on s, with length l, and gaps of d. Used to align multiple arrows.
}
\examples{
gt_shift()
}
\keyword{shift}
