% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_sss.R
\name{gt_sss}
\alias{gt_sss}
\title{Stochastic stabiity}
\usage{
gt_sss(mat = gt_sss_mat, e = 1, t = 1, cex.pt = 8, ann = TRUE,
  mar = 1.5, oma = 5, bgcol = "gray90", pointcol = "gray50",
  labcex = 0.5, indent = 4, arrows = TRUE, etlabs = TRUE,
  outax = TRUE)
}
\arguments{
\item{e}{Error}

\item{t}{time}
}
\description{
Plot matrices for different values of e and t. Thanks to Bernd Beber for generating Much of this code.
}
\examples{
gt_sss()
}
\keyword{Stability}
\keyword{Stochastic}

