% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_star.R
\name{gt_star}
\alias{gt_star}
\title{Helper to draw stars}
\usage{
gt_star(x, y, rad = 1, phi = 0, starfill = "turquoise", outer = "black",
  tips = 8, starborderlwd = 3)
}
\arguments{
\item{tips}{number of points}
}
\description{
A function to draw stars
}
\examples{
gt_star()

}
\keyword{stars}

