% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{gt_BRarrow}
\alias{gt_BRarrow}
\title{Helper to make best response arrows for bimatrix}
\usage{
gt_BRarrow(X, Y, vert = FALSE, arrowdist = 0.2, space = 0.15,
  color = gray(0.3), nash = TRUE, width = 3, alength = 0.25)
}
\arguments{
\item{X}{x coordinated for arrow}

\item{Y}{y coordinated for arrow}
}
\description{
Helper to make best response arrows for bimatrix
}
\examples{
gt_BRarrow()
}
\keyword{arrow}
