% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_brgraph.R
\name{gt_brgraph}
\alias{gt_brgraph}
\title{Plot best reactions for two person normal form games}
\usage{
gt_brgraph(R, C = t(R), PR = "Row player", PC = "Column Player",
  labelsR = c("C", "D"), labelsC = labelsR, legBR = TRUE, legr = TRUE,
  legc = TRUE, leglocBR = c(0.8, 0.2), legxyr = c(c(0.8, 3 * (min(C) +
  max(C))/4)), legxyc = c(c(0.8, (min(C) + max(C))/4)), ulegend = TRUE,
  BR = TRUE, draw1 = TRUE, draw2 = TRUE, vert = TRUE, util = FALSE,
  Rloc = NULL, Cloc = NULL, ur = TRUE, uc = TRUE, cthighlight = FALSE,
  rthighlight = FALSE, cbhighlight = FALSE, rbhighlight = FALSE,
  mmcol = "turquoise", xlabu = paste("Probability on", labelsC[1], "for",
  PC), ylabu = paste("Utility for", PR), legtextr = c(paste("u(",
  labelsR[1], ")", sep = ""), paste("u(", labelsR[2], ")", sep = "")),
  legtextc = c(paste("u(", labelsC[2], ")", sep = ""), paste("u(", labelsC[1],
  ")", sep = "")), mainbr = "", mainu = "", colown = "red",
  coloth = gray(0.4), hl_lwd = 8, lwd = 2)
}
\arguments{
\item{R}{Row payoffs}

\item{C}{Column payoffs}
}
\description{
This function plots best reactions. Intersections are Nash equilibria. Thanks to Neelan Sircar for developing much of this code.
To show utilities for players use \code{util = TRUE}. To show best response functions use \code{BR = TRUE}. Both can be done also
with output either aligned horizontally or vertically, using \code{vert}.
}
\examples{
C <- matrix(c(1,0,0,1),2,2)
R <- matrix(c(1,0,0,1),2,2)
gt_brgraph(R, C, PR="Country R", PC="Country C", leglocBR=c(.8,.2))
}
\keyword{Best}
\keyword{Nash,}
\keyword{Payoff}
\keyword{matrix,}
\keyword{responses}

