% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_declarations.R
\name{gt_declarations}
\alias{gt_declarations}
\title{# Legislative bargaining with three equally weighted voters. For any recognition probabilities, return the equilibrium allocations. Note: when lowest unit has 0 probability then a random solution taken}
\usage{
gt_declarations(x)
}
\arguments{
\item{x}{Vector with signals received}
}
\description{
# Legislative bargaining with three equally weighted voters. For any recognition probabilities, return the equilibrium allocations. Note: when lowest unit has 0 probability then a random solution taken
}
\examples{
gt_declarations(c(1,0,1,1,0))
 gt_declarations(c(0,0,1,1,0))
}
\keyword{Cascade}

