% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_folkmm.R
\name{gt_folkmm}
\alias{gt_folkmm}
\title{Graphs the individually rational and feasible outcomes that can be sustained as equilibria of a two person game}
\usage{
gt_folkmm(X, Y = t(X), fine = 100, feasible = TRUE, rational = TRUE,
  pointsize = 2, colf = rgb(0, 1, 0, 0.5), colr = rgb(0.5, 0, 0.5, 0.5),
  main = "", mainsize = 1, ...)
}
\arguments{
\item{X}{Payoff matrix for player 1. Defaults to coordination.}

\item{Y}{Payoff matrix for player 2. Defaults to coordination.}
}
\description{
Graphs the individually rational and feasible outcomes that can be sustained as equilibria of a two person game
}
\examples{
X = matrix(c(2,3,0,1),2)
gt.folkmm(X)
X = matrix(  c(2,0,0,0,1,0,0,0,5),3)
Y = t(matrix(c(1,2,1,2,1,1,1,3,1),3))
gt.folkmm(X,Y)
}
\keyword{Nash}
\keyword{Payoff}
\keyword{matrix,}

