% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_jury.R
\name{gt_jury}
\alias{gt_jury}
\title{Condorcet jury theorem illustration}
\usage{
gt_jury(n, fill = TRUE, bgcol = "grey", fillcol = "red",
  fillcol2 = "white", base = 0.51, lwd = 1, xlim = c(-0.3, n + 0.3),
  ...)
}
\arguments{
\item{n}{number of voters}

\item{base}{probability each is right, defaults to 0.51}
}
\description{
Illustrates basic logic from law of large numbers.
}
\examples{
 gt_jury(20)

}
\keyword{Condorcet,}
\keyword{Jury}

