% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_nbs.R
\name{gt_nbs}
\alias{gt_nbs}
\title{Nash bargaining solution}
\usage{
gt_nbs(u1 = function(x) x^0.5, u2 = function(x) (1 - x),
  solution_only = FALSE, setcol = "red", flatcol = "grey", length = 500,
  xtop = 1, xlim = c(0, xtop), ylim = xlim, alt = NULL,
  altlab = FALSE, paretoset_border = NA, NBScol = "black",
  diagonal = TRUE, at = c(0, 1))
}
\arguments{
\item{u1}{a function, increasing in x}

\item{u2}{a function, decreasing in x}

\item{solution_only}{Just show solution}
}
\description{
Nash bargaining solution
}
\examples{
 gt_nbs(u1 = function(x) x^.5, u2 = function (x) 1 - x)
}
\keyword{Nash}
\keyword{bargaining}

