% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_banzhaf.R
\name{gt_plot_banzhaf}
\alias{gt_plot_banzhaf}
\title{Plot Banzhaf index against normalized raw weights}
\usage{
gt_plot_banzhaf(w, q, col1 = "red", col2 = "white")
}
\arguments{
\item{w}{Vector of voting weights}

\item{q}{Threshold for voting success}
}
\description{
Plot Banzhaf index against normalized raw weights
}
\examples{
 gt_plot_banzhaf(w = 1:5, q=.5)
 gt_plot_banzhaf(w=c(1, 1, 3,7,9,9)/30, q = .5)
}
\keyword{Banzhaf,}
\keyword{Decision}
\keyword{Rules}

