% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_shadowtext.R
\name{gt_shadowtext}
\alias{gt_shadowtext}
\title{Helper for shadow text}
\usage{
gt_shadowtext(x, y = NULL, labels, col = "black", bg = "white",
  n_shadows = 12, theta = seq(pi/4, 2 * pi, length.out = n_shadows),
  r = 0.1, ...)
}
\arguments{
\item{x}{X location}

\item{y}{Y location}

\item{labels}{text}

\item{col}{Text color}

\item{bg}{Background (shadow) color}

\item{n_shadows}{Number of shadows to cast: deault is 12}
}
\description{
Place a shadow behind text (white default)
Helper to make text more visible
Credit: Greg Snow http://blog.revolutionanalytics.com/2009/05/make-text-stand-out-with-outlines.html
}
\examples{
frame()
gt_shadowtext(x = .5, y = .5, labels = "text", col = "white", bg = "red")
}
\keyword{text}

