% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_sss.R
\name{gt_sss_mat}
\alias{gt_sss_mat}
\title{Stochastic stability}
\usage{
gt_sss_mat(e = 0.1)
}
\arguments{
\item{e}{Error}
}
\description{
Generate transition matrix for Stochastic stability demonstration. Takes error \code{e} as argument
}
\examples{
gt_sss_mat()
}
\keyword{Stability}
\keyword{Stochastic}

